keiba_ai/requirements.txt
# 競馬予測AIシステム 必要パッケージ一覧
# Python 3.11.4推奨

# データ操作
pandas>=2.2.0
numpy>=1.24.0

# スクレイピング
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0
playwright>=1.40.0

# 機械学習
lightgbm>=4.1.0
scikit-learn>=1.3.0

# 可視化
matplotlib>=3.8.0
seaborn>=0.13.0
japanize-matplotlib>=1.1.3

# ユーティリティ
tqdm>=4.66.0
pyyaml>=6.0.0
apscheduler>=3.10.0

# HTMLパース
html5lib>=1.1

# その他
openpyxl>=3.1.0
xlrd>=2.0.0
